LayoutTests/http/tests/loading/preload-script-src-with-csp-nonce-expected.txt

main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This test makes sure that the Preload scanner preloads a script resource with a nonce attribute.
PASS
